feat: 아티스트 관련 API 통합 및 불필요한 파일 제거
